src/backend/pyproject.toml
[project]
name = "liveness-backend"
version = "0.1.0"
description = "FastAPI backend for real-time liveness detection"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Cataldir" }]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "liveness-detector>=0.0.6; platform_system != 'Windows'",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.10.0",
    "scikit-learn>=1.7.2",
    "azure-ai-projects>=1.0.0",
    "azure-identity>=1.25.1",
    "openai>=2.8.1",
    "azure-ai-vision-face>=1.0.0b2",
]

[tool.uv]
dev-dependencies = [
    "black>=24.4.0",
    "pylint>=3.2.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
